DXE-3418 change ExpectedTTL type in DNSSEC to int64

+++ pkg/dns/zone.go
@@ -213,7 +213,7 @@ type (
 	SecRecords struct {
 		DNSKeyRecord     string    `json:"dnskeyRecord"`
 		DSRecord         string    `json:"dsRecord"`
-		ExpectedTTL      int       `json:"expectedTtl"`
+		ExpectedTTL      int64     `json:"expectedTtl"`
 		LastModifiedDate time.Time `json:"lastModifiedDate"`
 	}
 )
